\name{engelberg.annmax}
\alias{engelberg.annmax}
\docType{data}
\title{ Example datasets for gev }
\description{
  Yearly Maxima of daily precipitation at Engelberg Switzerland.
  Data are in mm.

}
\usage{
  data(engelberg.annmax)
}
\format{
  A data frame with two columns: First column (MAX) is the precipitation 
  maximum in the year, Second column (lab) the day when the maximum was
  reached in that year. Format of the date is "yyyy.mm.dd".
}
\details{

}
\source{
  Data from MeteoSwiss Zurich, Switzerland
}
\references{
  
}
\examples{
data(engelberg.annmax)
engelberg.annmax

}
\keyword{datasets}
